contact_management_system/requirements.txt
Django==3.0.14
djangorestframework==3.14.0